# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "运维发布管理系统 API 文档",
        "title": "OpsCenter API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@opscenter.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:18080",
    "basePath": "/api",
    "paths": {
...
        "/servers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取服务器列表，可按类型和启用状态筛选",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器"
                ],
                "summary": "获取服务器列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务器类型 (lvs/k8s/nginx/preprod)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "是否显示全部 (true/false)",
                        "name": "all",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ServerResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新服务器，仅管理员可用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器"
                ],
                "summary": "创建服务器",
                "parameters": [
                    {
                        "description": "服务器信息",
                        "name": "server",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Server"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.ServerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/servers/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据 ID 获取服务器详情（脱敏）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器"
                ],
                "summary": "获取服务器详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务器 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ServerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新服务器信息，仅管理员可用。密码字段传 __keep__ 表示保留原值",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器"
                ],
                "summary": "更新服务器",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务器 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "服务器信息",
                        "name": "server",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Server"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ServerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除指定服务器（软删除），仅管理员可用",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器"
                ],
                "summary": "删除服务器",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务器 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/servers/{id}/edit": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取服务器完整信息（密码用掩码显示），仅管理员可用",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器"
                ],
                "summary": "获取服务器编辑信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务器 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/servers/{id}/test": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "测试 SSH 连接到指定服务器，仅管理员可用",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器"
                ],
                "summary": "测试服务器连接",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务器 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/servers/{id}/toggle": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "启用或禁用服务器，仅管理员可用",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器"
                ],
                "summary": "切换服务器启用状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务器 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Server": {
            "type": "object",
            "properties": {
                "auth_type": {
                    "type": "string"
                },
                "backup_path": {
                    "type": "string"
                },
                "config_path": {
                    "type": "string"
                },
                "config_pattern": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "env": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "private_key": {
                    "type": "string"
                },
                "script_password": {
                    "type": "string"
                },
                "script_path": {
                    "type": "string"
                },
                "server_type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.ServerResponse": {
            "type": "object",
            "properties": {
                "auth_type": {
                    "type": "string"
                },
                "backup_path": {
                    "type": "string"
                },
                "config_path": {
                    "type": "string"
                },
                "config_pattern": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "env": {
                    "type": "string"
                },
                "has_password": {
                    "type": "boolean"
                },
                "has_private_key": {
                    "type": "boolean"
                },
                "has_script_password": {
                    "type": "boolean"
                },
                "host": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "script_path": {
                    "type": "string"
                },
                "server_type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}